+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Collect and show windows device information.
+    - JavaVersioncheckResult
+
 ## [1.4.0] - 2023-12-23
 
 ### Added